add list command and error check

+++ gadget-lib.sh
@@ -77,7 +77,7 @@ function create_mass_storage() {
         echo 1 > ${FUNCTION_DIR}/lun.0/nofua
         ;;
       *)
-        echo "invalid option: $1"
+        echo "invalid option: $1" >&2
         ;;
     esac
     shift
@@ -116,15 +116,25 @@ function remove_function () {
   GADGET_DIR="${CONFIGFS_USB_GADGET}/${GADGET_NAME}"
   FUNCTION_DIR="${GADGET_DIR}/functions/${FUNCTION_TYPE_NAME}"
 
+  if [ ! -d ${FUNCTION_DIR} ]; then
+    echo "Invalid function name: ${FUNCTION_TYPE_NAME}" >&2
+    return 1
+  fi
+
   rm -f "${GADGET_DIR}/configs/c.1/${FUNCTION_TYPE_NAME}"
   rmdir "${FUNCTION_DIR}"
 
-  return
+  return 0
 }
 
 function create_gadget () {
   GADGET_NAME="$1"
   GADGET_DIR="${CONFIGFS_USB_GADGET}/${GADGET_NAME}"
+
+  if [ -d ${GADGET_DIR} ]; then
+    echo "Already exists: ${GADGET_NAME}" >&2
+    return 1
+  fi
 
   mkdir -p ${GADGET_DIR}
 
@@ -154,10 +164,26 @@ function remove_gadget () {
   GADGET_DIR="${CONFIGFS_USB_GADGET}/${GADGET_NAME}"
   CONFIG_DIR="${GADGET_DIR}/configs/c.1"
 
-  disable_gadget "${GADGET_NAME}"
+  if [ ! -d ${GADGET_DIR} ]; then
+    echo "Invalid gadget name: ${GADGET_NAME}" >&2
+    return 1
+  fi
 
-  rmdir ${CONFIG_DIR}
-  rmdir ${GADGET_DIR}/strings/*
+  if [ ! -z $(list_gadget ${GADGET_NAME}) ]; then
+    echo "Function is existing" >&2
+    return 2
+  fi
+
+  disable_gadget ${GADGET_NAME}
+
+  if [ -d ${CONFIG_DIR} ]; then
+    rmdir ${CONFIG_DIR} || die
+  fi
+
+  if [ -d ${GADGET_DIR}/strings ]; then
+    rmdir ${GADGET_DIR}/strings/* || die
+  fi
+
   rmdir ${GADGET_DIR}
 
   return
@@ -166,6 +192,11 @@ function remove_gadget () {
 function enable_gadget () {
   GADGET_NAME="$1"
   GADGET_DIR="${CONFIGFS_USB_GADGET}/${GADGET_NAME}"
+
+  if [ ! -z $(cat ${GADGET_DIR}/UDC) ]; then
+    echo "Gadget is already enabled" >&2
+    return
+  fi
 
   ls /sys/class/udc > ${GADGET_DIR}/UDC
 
@@ -176,7 +207,24 @@ function disable_gadget () {
   GADGET_NAME="$1"
   GADGET_DIR="${CONFIGFS_USB_GADGET}/${GADGET_NAME}"
 
+  if [ -z $(cat ${GADGET_DIR}/UDC) ]; then
+    return
+  fi
+
   echo "" > ${GADGET_DIR}/UDC
+
+  return
+}
+
+function list_gadget () {
+  GADGET_NAME="$1"
+  GADGET_DIR="${CONFIGFS_USB_GADGET}/${GADGET_NAME}"
+
+  if [ ! -d "${GADGET_DIR}/functions" ]; then
+    return
+  fi
+
+  ls /sys/kernel/config/usb_gadget/g0/functions | grep -e ".*\..*" 2> /dev/null
 
   return
 }
